requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
cityscapesScripts==2.2.0
kornia==0.5.8
matplotlib==3.4.2
mmcv_full==1.3.7
numpy==1.19.2
Pillow==9.2.0
prettytable==2.1.0
seaborn==0.11.1
setproctitle==1.2.3
timm==0.3.2
tensorboard
packaging==21.3
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu110
torchvision==0.8.2+cu110